refactor(mcp): converge runtime lifecycle ownership
